+++ src/components/layout/infoMenuItem/InfoMenuActivity.vue
